pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app/settings.py"]

[project]
name = "voicetovoic"
version = "0.0"
dependencies = [
    "aiohttp[speedups]>=3.11.11",
    "fastapi[standard]>=0.115.6",
    "loguru>=0.7.3",
    "openai>=1.59.8",
    "psycopg[binary,pool]>=3.2.3",
    "pydantic-ai-slim[groq]>=0.0.19",
    "pydantic-settings>=2.7.1",
    "websockets>=14.1",
    
]
requires-python = ">=3.12.8"